fix: resolve circular dependency in thread runs
