+++ src/leapfrogai_api/routers/openai/runs.py
@@ -5,14 +5,8 @@ from fastapi import HTTPException, APIRouter, status
 from fastapi.responses import StreamingResponse
 from openai.types.beta.threads import Run
 from openai.pagination import SyncCursorPage
-from leapfrogai_api.backend.types import (
-    ModifyRunRequest,
-)
 from leapfrogai_api.routers.openai.requests.thread_run_create_params_request import (
     ThreadRunCreateParamsRequest,
-)
-from leapfrogai_api.typedef.request import (
-    RunCreateParamsRequest,
 )
 from leapfrogai_api.data.crud_run import CRUDRun
 from leapfrogai_api.data.crud_thread import CRUDThread
@@ -21,6 +15,8 @@ from leapfrogai_api.utils.validate_tools import (
     validate_assistant_tool,
     validate_assistant_tool_choice_option,
 )
+from leapfrogai_api.typedef.request import RunCreateParamsRequest, ModifyRunRequest
+from leapfrogai_api.backend.thread_runner import ThreadRunner
 
 router = APIRouter(prefix="/openai/v1/threads", tags=["openai/threads/runs"])
 
@@ -60,7 +56,12 @@ async def create_run(
         )
 
     try:
-        return await request.generate_response(existing_thread, new_run, session)
+        return await ThreadRunner().generate_response(
+            request=request,
+            existing_thread=existing_thread,
+            new_run=new_run,
+            session=session,
+        )
     except Exception as exc:
         traceback.print_exc()
         raise HTTPException(
@@ -84,7 +85,7 @@ async def create_thread_and_run(
         )
 
     try:
-        return await request.generate_response(new_run, new_thread, session)
+        return await ThreadRunner().generate_response(new_run, new_thread, session)
     except Exception as exc:
         traceback.print_exc()
         raise HTTPException(
