+++ src/leapfrogai_api/typedef/__init__.py
@@ -1,1 +1,2 @@
 from .model_types import Model as Model
+from .assistant_types import ListAssistantsResponse as ListAssistantsResponse
