+++ src/leapfrogai_api/typedef/request/__init__.py
@@ -6,3 +6,10 @@ from .run_create_params_base import (
     RunCreateParamsRequestBase as RunCreateParamsRequestBase,
 )
 from .run_create_params import RunCreateParamsRequest as RunCreateParamsRequest
+from .run_modify import ModifyRunRequest as ModifyRunRequest
+from .audio_types import (
+    CreateTranscriptionRequest as CreateTranscriptionRequest,
+    CreateTranscriptionResponse as CreateTranscriptionResponse,
+    CreateTranslationRequest as CreateTranslationRequest,
+    CreateTranslationResponse as CreateTranslationResponse,
+)
