+++ src/leapfrogai_api/typedef/request/run_modify.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel, Field
+
+
+class ModifyRunRequest(BaseModel):
+    """Request object for modifying a run."""
+
+    metadata: dict[str, str] | None = Field(default=None, examples=[{}])
